Fixing permissions (issue 17) finished S5,S6

+++ trunk/www/app/controllers/Application.java
@@ -19,38 +19,50 @@ import play.mvc.With;
  * 
  * @author mahmoudsakr
  */
-@With (Secure.class)
-public class Application extends SmartController {
-	public static String hash(String str) {
+@With( Secure.class )
+public class Application extends SmartController
+{
+	public static String hash( String str )
+	{
 		String res = "";
-		try {
-			MessageDigest algorithm = MessageDigest.getInstance("MD5");
+		try
+		{
+			MessageDigest algorithm = MessageDigest.getInstance( "MD5" );
 			algorithm.reset();
-			algorithm.update(str.getBytes());
+			algorithm.update( str.getBytes() );
 			byte[] md5 = algorithm.digest();
 			String tmp = "";
-			for (int i = 0; i < md5.length; i++) {
-				tmp = (Integer.toHexString(0xFF & md5[i]));
-				if (tmp.length() == 1) {
+			for( int i = 0; i < md5.length; i++ )
+			{
+				tmp = (Integer.toHexString( 0xFF & md5[i] ));
+				if( tmp.length() == 1 )
+				{
 					res += "0" + tmp;
-				} else {
+				}
+				else
+				{
 					res += tmp;
 				}
 			}
-		} catch (NoSuchAlgorithmException ex) {
+		}
+		catch( NoSuchAlgorithmException ex )
+		{
 		}
 		return res;
 	}
 
-	public static String randomHash() {
-		return randomHash(32);
+	public static String randomHash()
+	{
+		return randomHash( 32 );
 	}
 
-	public static String randomHash(int length) {
-		return hash(System.currentTimeMillis() * Math.random() + "").substring(0, length);
+	public static String randomHash( int length )
+	{
+		return hash( System.currentTimeMillis() * Math.random() + "" ).substring( 0, length );
 	}
 
-	public static void index() {
+	public static void index()
+	{
 		render();
 	}
 
@@ -62,14 +74,15 @@ public class Application extends SmartController {
 	 * @author Amr Hany
 	 * @param ProjectID
 	 */
-	public static void viewComponents(long id) {
+	public static void viewComponents( long id )
+	{
 
-		Project currentProject = Project.findById(id);
-		boolean inSprint = (currentProject.inSprint(new Date()));
+		Project currentProject = Project.findById( id );
+		boolean inSprint = (currentProject.inSprint( new Date() ));
 		String projectName = currentProject.name;
-		List<Component> components = Component.find("byProject.idAnddeleted", id, false).fetch();
+		List<Component> components = Component.find( "byProject.idAnddeleted", id, false ).fetch();
 
-		render(components, id, projectName, inSprint);
+		render( components, id, projectName, inSprint, currentProject );
 
 	}
 
@@ -80,11 +93,12 @@ public class Application extends SmartController {
 	 * @author Amr Hany
 	 * @param componentID
 	 */
-	public static void viewComponent(long id) {
+	public static void viewComponent( long id )
+	{
 
-		Component component = Component.findById(id);
-		boolean inSprint = component.project.inSprint(new Date());
-		render(component, inSprint);
+		Component component = Component.findById( id );
+		boolean inSprint = component.project.inSprint( new Date() );
+		render( component, inSprint );
 	}
 
 	/**
@@ -94,25 +108,33 @@ public class Application extends SmartController {
 	 * @param componentID
 	 */
 
-	@Check ("canDeleteComponent")
-	public static void deleteComponent(long id) {
-		Component c = Component.findById(id);
-		c.deleteComponent();
-		Logs.addLog(Security.getConnected(), "Delete", "Component", c.id, c.project, new Date(System.currentTimeMillis()));
-
+	// @Check( "canDeleteComponent" )
+	public static void deleteComponent( long id )
+	{
+		Component c = Component.findById( id );
+		if( Security.getConnected().in( c.project ).can( "deleteComponent" ) )
+		{
+			c.deleteComponent();
+			Logs.addLog( Security.getConnected(), "Delete", "Component", c.id, c.project, new Date( System.currentTimeMillis() ) );
+		}
+		else
+			forbidden();
 	}
 
-	public static void md5(String str) {
-		renderText(hash(str));
+	public static void md5( String str )
+	{
+		renderText( hash( str ) );
 	}
 
-	@Check ("systemAdmin")
-	public static void adminIndexPage() {
+	@Check( "systemAdmin" )
+	public static void adminIndexPage()
+	{
 		render();
 	}
 
-	@Check ("systemAdmin")
-	public static void adminIndex() {
+	@Check( "systemAdmin" )
+	public static void adminIndex()
+	{
 		render();
 	}
 
@@ -122,13 +144,15 @@ public class Application extends SmartController {
 	 * @param id
 	 *            user id
 	 */
-	@Check ("canEditProfile")
-	public static void profile(long id) {
-		if (id == 0) {
+	@Check( "canEditProfile" )
+	public static void profile( long id )
+	{
+		if( id == 0 )
+		{
 			id = Security.getConnected().id;
 		}
-		User user = User.findById(id);
-		render(user);
+		User user = User.findById( id );
+		render( user );
 	}
 
 	/**
@@ -141,35 +165,40 @@ public class Application extends SmartController {
 	 * @param id
 	 *            user id
 	 */
-	@Check ("canEditProfile")
-	public static void editProfile(@Required (message = "You must enter a name") String name, String pwd1, String pwd2, @Required (message = "You must enter an email") @Email (message = "You must enter a valid email") String email, long id) {
-		if (Validation.hasErrors() || (pwd1.length() > 0 && !pwd1.equals(pwd2))) {
-			flash.error("An error has occured");
-			profile(id);
+	@Check( "canEditProfile" )
+	public static void editProfile( @Required( message = "You must enter a name" ) String name, String pwd1, String pwd2, @Required( message = "You must enter an email" ) @Email( message = "You must enter a valid email" ) String email, long id )
+	{
+		if( Validation.hasErrors() || (pwd1.length() > 0 && !pwd1.equals( pwd2 )) )
+		{
+			flash.error( "An error has occured" );
+			profile( id );
 		}
 
-		User usr = User.findById(id);
+		User usr = User.findById( id );
 		String oldEmail = usr.email;
 		usr.name = name;
-		if (pwd1.length() > 0)
-			usr.pwdHash = Application.hash(pwd1);
+		if( pwd1.length() > 0 )
+			usr.pwdHash = Application.hash( pwd1 );
 		usr.email = email;
 		usr.save();
 		// Added By Wallas in Sprint 2.
-		if (!usr.email.equals(oldEmail)) {
-			usr.activationHash = Application.randomHash(32);
+		if( !usr.email.equals( oldEmail ) )
+		{
+			usr.activationHash = Application.randomHash( 32 );
 			usr.isActivated = false;
 			usr.save();
-			session.put("username", email); // Update the session cookie by
+			session.put( "username", email ); // Update the session cookie by
 			// setting the new Email.
 			String subject = "Your SmartSoft new Email activation requires your attention";
 			String body = "Dear " + usr.name + ", You have requested to change the Email Address associated with your account. Please click the following link to activate your account: " + "http://localhost:9000/accounts/doActivation?hash=" + usr.activationHash;
-			Mail.send("se.smartsoft@gmail.com", usr.email, subject, body);
-			flash.success("Successfully saved your data! , please check your new Email and follow the instructions sent by us to confirm your new Email.");
-			profile(id);
-		} else {
-			flash.success("Successfully saved your data!");
-			profile(id);
+			Mail.send( "se.smartsoft@gmail.com", usr.email, subject, body );
+			flash.success( "Successfully saved your data! , please check your new Email and follow the instructions sent by us to confirm your new Email." );
+			profile( id );
+		}
+		else
+		{
+			flash.success( "Successfully saved your data!" );
+			profile( id );
 		}
 
 	}
